Fix main_test, add startup logic to splash_screen
